refactor (cli_utils): moved get_sample_name and get_analysis_dir to report
