+++ BALSAMIC/tools/__init__.py
@@ -1,4 +1,4 @@
 from .rule_utils import (get_sample_type, get_vcf, get_result_dir,
                          get_conda_env, get_picard_mrkdup, get_chrom)
 from .cli_utils import (add_doc, createDir, get_packages, get_package_split,
-                        get_sample_name, get_ref_path, iterdict)
+                        get_ref_path, iterdict)
