+++ BALSAMIC/tools/cli_utils.py
@@ -151,36 +151,6 @@ def get_package_split(conda_yamls):
     return (pkgs)
 
 
-def get_sample_name(json_in):
-    """
-    Get sample name from input json file
-    """
-
-    try:
-        with open(os.path.abspath(json_in), "r") as fn:
-            sample_json = json.load(fn)
-            sample_name = sample_json["analysis"]["sample_id"]
-    except OSError:
-        print("Couldn't load json file or file path is not absolute")
-
-    return sample_name
-
-
-def get_analysis_dir(json_in, dir_type):
-    """
-    Get analysis dir from input json file
-    """
-
-    try:
-        with open(os.path.abspath(json_in), "r") as fn:
-            sample_json = json.load(fn)
-            analysis_dir = sample_json["analysis"][dir_type]
-    except OSError:
-        print("Couldn't load json file or file path is not absolute")
-
-    return analysis_dir
-
-
 def get_ref_path(reference_config):
     """
     Set full path to reference files
